backend/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8,<=3.0
django-cors-headers>=3.7.0,<=3.8.0
django-filter>=2.4.0,<=2.5.0
uWSGI>=2.0.0,<=2.1.0
Pillow>=9.1.0,<=9.2.0
martor>=1.6.13,<=1.7.0
django-light==0.1.0.post3
djangorestframework>=3.13.1,<=3.13.9
drf-yasg>=1.20.0,<=1.21.0